Temporarily disable add / edit rule page
The page currently defaults to adding a number of dangerous resources
like OpenShift `oauthaccesstokens` by default under the "Safe Resources"
section. This is because the list of admin resources is hard-coded and
unknown resources are assumed to be "safe." This became a problem when
we switched to API discovery, which added a bunch of new resources.

We should revisit the approach we take to building the safe resources
list.