Monitoring: Fix handling of Alert state changes on details page
Fix the Alert details page to correctly handle the case where an Alert
changes from Not Firing to Firing or Pending while it is being viewed.
Previously, it would switch to displaying the Firing or Pending Alert,
but that doesn't work if there are multiple Alerts for the Rule. Now the
page just updates to reflect the state change.

Also improve handling of Alerts that switch to "Not Firing" while being
viewed. Previously, the page could change to a "Not Found" message
depending on how the labels changed. Now it just switches to showing the
Alert as "Not Firing".