Bug 1656730 - Remove `customresourcedefinitions` from CRD paths
The original change was put in place to keep the CRDs nav item
highlighted, but I'm not sure the new behavior is better.

* It introduced some problems switching between projects when a CRD is
  selected.
* It created two different URL patterns to the same resource list. You
  might get one or the other depending on how you get to the list in the
  UI. The CRDs nav item is highlighted or not depending on where you
  came from.
* It prevented you from seeing the custom list columns when viewing a
  resource. This was originally listed as a benefit, but I'd argue in
  most cases you want the custom columns. I personally prefer having
  one consistent view of the resource no matter where it shows up
  (search view, CRDs page, or another page).

This change does have the drawback of not highlighting the CRDs nav item. I
think the behavior is fine though since you're no longer looking at the
CRDs themselves.

https://bugzilla.redhat.com/show_bug.cgi?id=1656730