Firehose: honor `optional` prop for watches on individual resources (#1711)
* Firehose: introduce "ignoreIfMissing" resource property

If set to `true`, a resource is skipped from generic
`loadError` 404 checking (cross-resources). So the overall data
retrieval does not fail if a specified resource is missing.

Can be used to optionally query a resource but enable rendering
of children if it is missing (HTTP 404).

The new property is an addition to existing `optional` property which
does not block rendering of children if list-retrieval is still in progress.

With default value `ignoreIfMissing = false`, the component is
backward compatible.

* Firehose: fix the `optional` property for 404

If a resource is `optional: true` then it is excluded from
setting of top-level loadError.

Firehose's child component can still check loadError by
inspecting the resource directly.