Query Browser: Fix display of range vector queries in results table
Instead of starting with the `query_range` API result and augmenting it
with the `query` API result, start with the `query` result and augment
with the `query_range` results.

Change to display `query` API error messages, not just `query_range`
error messages.

Add missing `aria-label` to `Table` and remove `key` from `TableBody`,
which forced re-renders.