Bug 1760376: Update create operand breadcrumb
Go to the ClusterServiceVersion Overview tab instead of the resource tab
when clicking on the breadcrumb of the create operand form. Often,
you've arrived at the create operand form by clicking on "Create
Instance" in one of the cards. Returning to the resource tab instead of
Overview can be disorienting.