Lazy load images in OperatorHub
This addresses some performance problems we're seeing loading images in
OperatorHub. Specifically, the subscribe flow is blocked until all
images from the catalog load. Since we do code splitting, the request
for the JS chunk for the next page is blocked because of the pending
image requests. This has also caused timeouts in our OLM integration tests.