Query Browser: Fix "reduced resolution" message being shown briefly
When zooming out, the `Displaying with reduced resolution due to large
dataset` message was displayed briefly while the data updated. This
fixes that by disabling the message when `updating` is `true` and also
sets the `updating` flag in a `useLayoutEffect` hook instead of a
`useEffect` hook so that it gets set before the changes are painted to
the screen.