Monitoring Dashboards: Tooltips can overflow graph, move above cursor
Allow tooltips to overflow the graph container.

On the Administrator perspective Monitoring Dashboards page, move the
tooltip above the mouse cursor. This stops it overlapping with the graph
legend or with any graph below. This works around a problem where those
SVG elements were visible through the tooltip. However, if there is
another graph to the right, its elements can still overlap and be
visible through the tooltip.