Specify the destination of bugs & support cases
This adds verbiage to clarify that bugs and support cases will be opened with Red Hat (which was implicit before). We have seen a few instances of customers opening support cases with Red Hat for IBM products because the link was present in the console; by adding this & working with IBM to leverage the ConsoleLink CRD to add a "Open Cloud Pak Support Case with IBM" link to help provide an alternative option.