Replace deprecated Formik render function with child callback function.
This removes a technical deps and allow us to update Formik in the
future. It also removes this warning when the user opens the screen:

Warning: <Formik render> has been deprecated and will be removed in
future versions of Formik. Please use a child callback function
instead. To get rid of this warning,
replace <Formik render={(props) => ...} />
with <Formik>{(props) => ...}</Formik>