Add Thanos tenancy proxy for rules
The Thanos querier deployment exposes a tenancy-aware service for
recording and alerting rules. It is exposed via a different port than
the existing tenancy-aware service for queries because it requires
different permissions to be granted.

This change modifies the console bridge to make this service available
to the console users.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>