Bug 1965391: Skip to 3rd step when SC present
Jump to the 3rd step in attached-devices mode when a no-provisioner
storage class has already been created. Also, keep the URL and wizard in
sync while doing so, for an overall better user experience (for
instance, if the user wants to bookmark the page they are currently on,
hence I felt that synchronization was necessary).

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>