Monitoring: Fix `Created by` field when refreshing Create Silence form
If you went to the Create Silence form and then refreshed the page, the
form would load before the `user` object was populated in Redux, causing
the "Created by" field to not be populated. This fixes that by ensuring
that the form responds when the `user` object is updated in Redux.