Bug 2008256: Update Internationalization README with scope info
No scope information was provided in the internationalization README. To know scope information, you needed to go back to old documents or read old Slack threads that may not be available at this point. We should add this information to make it easier for the team to identify bugs that are out of scope at this time.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2008256