Monitoring dashboards: Improve series labels ordering in tooltips
The number of labels is often too great to fit in the tooltip and
sometimes the first one or two labels have the same value for all
series, making the text in the tooltip useless. This tries to order the
labels intelligently so that those with different values are visible.