Add new option for VIRTIOWIN ConfigMap name
In OpenShift Virtualization 4.10.0, the VM-Import operator has been removed from the package, along with the  config map.
However, virtio-win image pullspec, which was included in that config map, is still required by the UI and has nothing to do with V2V.
We are re-introducing this ConfigMap, but with more accurate name - .
This PR adds one more option for that config map name, to support a smooth upgrade path, so it will work both with OCP 4.10 + CNV 4.9 and OCP 4.10 + CNV 4.10.

Signed-off-by: orenc1 <ocohen@redhat.com>