Bug 2106866: Fix flaky OLM descriptor test
Use `{force: true}` on some of the click actions that are frequently flaking. There are race conditions between react re-rendering and cypress querying and clicking these elements. If a re-render occurs at exactly the right moment, the element cypress queried with `get` can be detached from the DOM, which causes `click` to fail. Forcing the click is not ideal, but will likely be less flaky. We stil need to do more investigation as to why these OLM pages are re-rendering frequently enough to cause these problems, but this is a stop-gap to improve CI pass rates.