frontend: fix kube-apiserver availability query
The original kube-apiserver query used to report 100% availability even
when a node was down whilst other control-plane components reported
66%. This is caused by the fact that the query relied on the number of
Prometheus targets of the kube-apiserver. However, when a node is down,
there is a controller in Kubernetes that removes the kube-apiserver
address on that node from the Endpoint that Prometheus scrapes. So
Prometheus is never able to detect scrape failures on that target since
it disappeared before becoming unavailable.

We found 3 different ways to resolve this issue:
1. Fix the upstream controller to keep the address around as a
   not ready address like it is done for the other components. This is
   by far the best solution, but this is a very fragile part of the
   Kubernetes codebase and the cost of such a change is enormous.
2. Change the Prometheus target for the kube-apiserver from the default
   Kubernetes Service to the kube-apiserver Service. This is a very nice
   and easy solution but it comes with one drawback which is that we
   won't monitor the endpoint that users are sending requests to.
3. Update the query to rely on the number of kube-apiserver pods instead
   of the Prometheus targets. This is also a nice and easy solution but
   this one doesn't seems to have any drawbacks, so that's the one we
   chose to go with after weighting cost vs value.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>