auth: implement session cookie per pod
Make it possible for multiple sessions to exist in parallel across
all the instances of the console.

This is achieved by naming the session cookies after the pod that
create them. That means that if there's a new rollout, a new cookie
will be created if browser traffic is redirected to a new console
instance.

The old cookies should eventually be removed once they reach their
MaxAge.