Show Observe section without PROMETHEUS and MONITORING flags
Currently, the Observe section is only shown to users which have the
`PROMETHEUS`, `MONITORING` and `CAN_GET_NS` flags.

Some dynamic plugins, for example distributed tracing and logging, add a
new link to the Observe section, which should be visible to all users -
even without the `PROMETHEUS`, `MONITORING` and `CAN_GET_NS` flags. Therefore this
PR removes the requirement on this flags.

I checked the links in the section, and the monitoring-related links
(Alerting, Metrics, Dashboards, Targets) set the required permissions
(`PROMETHEUS`, `MONITORING` and `CAN_GET_NS`) on the `console.navigation/href` already:
https://github.com/openshift/monitoring-plugin/blob/6b92084514ea9007d7d797f7699eab19e0c2f2fc/web/console-extensions.json

In case the Observe section is empty, it is hidden:
https://github.com/openshift/console/blob/e4887cb36842bd01b2958752aabeac556a9de50b/frontend/packages/console-app/src/components/nav/NavSection.tsx#L57-L60

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>